refactor : 커서 로직을 memberItems로 변경
