+++ src/main/java/com/programmers/bucketback/domains/bucket/repository/BucketRepositoryForCursor.java
@@ -2,7 +2,6 @@ package com.programmers.bucketback.domains.bucket.repository;
 
 import java.util.List;
 
-import com.programmers.bucketback.domains.bucket.application.vo.BucketMemberItemSummary;
 import com.programmers.bucketback.domains.bucket.application.vo.BucketSummary;
 import com.programmers.bucketback.domains.common.Hobby;
 
@@ -10,13 +9,6 @@ public interface BucketRepositoryForCursor {
 	List<BucketSummary> findAllByCursor(
 		final Long memberId,
 		final Hobby hobby,
-		final String cursorId,
-		final int pageSize
-	);
-
-	List<BucketMemberItemSummary> findBucketMemberItemsByCursor(
-		final List<Long> itemIdsFromBucketItem,
-		final Long memberId,
 		final String cursorId,
 		final int pageSize
 	);
