+++ src/main/java/com/programmers/bucketback/domains/item/repository/MemberItemRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.programmers.bucketback.domains.item.domain.Item;
 import com.programmers.bucketback.domains.item.domain.MemberItem;
 
-public interface MemberItemRepository extends JpaRepository<MemberItem, Long> {
+public interface MemberItemRepository extends JpaRepository<MemberItem, Long>, MemberItemRepositoryForCursor {
 
 	boolean existsMemberItemByMemberIdAndItem(Long memberId, Item item);
