+++ src/main/java/com/programmers/bucketback/domains/item/repository/MemberItemRepositoryForCursor.java
@@ -1,0 +1,15 @@
+package com.programmers.bucketback.domains.item.repository;
+
+import java.util.List;
+
+import com.programmers.bucketback.domains.bucket.application.vo.BucketMemberItemSummary;
+
+public interface MemberItemRepositoryForCursor {
+	List<BucketMemberItemSummary> findBucketMemberItemsByCursor(
+		final List<Long> itemIdsFromBucketItem,
+		final List<Long> itemIdsFromMemberItem,
+		final Long memberId,
+		final String cursorId,
+		final int pageSize
+	);
+}
